CLN: Split test_window.py
